Create README.md
Children In Need Token (CIN) is the world's first Crypto token designed to MAKE BIG CHANGE FOR SMALL PEOPLE - and give our investors an EVER INCREASING RETURN on investment. We are partnering with a select few children's charities around the world and our aim is to ensure that the people that need it most receive the funding directly. We intend to cut out any middle man so that every donation is used to its full capacity) We are here to make a difference but also recognise the need to ensure our investors are profitable. We see nothing wrong with these two working hand in hand. Through our unique "Tokenomics" our investors will have an ever growing portfolio and our ability to raise funds for worthy charities will only increase over time.